Changed test.ts and utils.ts to test.js, utils.js. Added all the helper function for tests in utils.js.
